New PlayerFinder methods for finding a player
Added 3 new methods to PlayerFinder for finding a player.
`find_with_name()` which looks for a player with a given name using the
MPRIS Identity value (case insensitive).
`wait_for_player()` and `wait_for_player_with_name()` which are blocking
versions of `find_active()` and `find_with_name()` respectively that block
until a player is found or a DBus error is encountered.
This is useful for when you want to wait for a player to be started but
don't want to constantly loop and look for a player.