Replace Duration and DurationExt
Instead of using `Duration` and working around the type conversions now
there's `MprisDuration` which will always be between 0 and i64::MAX
because the MPRIS spec returns length and position as a i64 but it can't
be negative. `MprisDuration` uses a u64 internally to avoid having to
deal with negative numbers when doing math operations. For Serde it acts
like a i64 since that's the type in the spec but it can be converted
to and from i64, u64 and Duration.