Rewrite the Metadata struct
The whole struct is now generated through the `gen_metadata_struct`
macro. This is done to avoid having to copy and paste the fields and
keys around while also making it easy to change the fields. The macro
generates a couple of methods and traits for the struct which are
documented on the macro itself.
Additionally `From<RawMetadata>` has been changed to `TryFrom` because
if there's a problem with the type the conversion should fail instead of
silently changing the values.